requirements.txt

Flask

flask_session

sentence-transformers

PyPDF2

python-docx

openai

pymongo

faiss-cpu

pandas

xlsxwriter

python-dotenv